% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayfit.R
\name{bayfit}
\alias{bayfit}
\title{Bayesian integration of a posterior}
\usage{
bayfit(lp, start, algorithm = "metropolis", options = list(), cl = NULL)
}
\arguments{
\item{lp}{Function that evaluates log-posterior given a vector of parameter values}

\item{start}{Initial values for the parameter values}

\item{algorithm}{A string indicating the MCMC algorithm to be used}

\item{options}{List of settings of the MCMC algorithm. See Details below.}

\item{cl}{Parallel cluster for running multiple chains in parallel}
}
\description{
Bayesian integration of a posterior
}
