% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{diagnostics}
\alias{diagnostics}
\title{Print technical information on the fitting of a model}
\usage{
diagnostics(x)
}
\arguments{
\item{x}{Model object belonging to the classes \code{\link{mlefit}}, \code{bayfit} or \code{nlm}.}
}
\description{
Print technical information on the fitting of a model
}
